First Progress Made
Started building the game by following the tutorials on the course. Have a basic tile map setup with prefabs that can easily be dragged and dropped. A labelling system that gives the coordinates of the tile. Finally, a basic system that gets a enemy object to move along the path.